supafastllm/docker/fastapi/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
aiofiles==23.2.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 